exchangerate/requirements.txt
google-cloud-bigquery
google-cloud-storage
requests
